feat: category update and delete secure

+++ src/services/category.service.ts
@@ -1,12 +1,35 @@
 import { prisma } from '../config/db'
 
-import { BadRequest, Conflict, CustomError, NotFound } from '../errors'
+import {
+  BadRequest,
+  Conflict,
+  CustomError,
+  NotAuthorized,
+  NotFound,
+} from '../errors'
 import { WorkspaceRepository, CategoryRepository } from '../core/repositories'
-import { CategoriesDto, CreateCategoryDto } from '../core/dtos'
+import {
+  CategoriesDto,
+  CreateCategoryDto,
+  UpdateCategoryDto,
+} from '../core/dtos'
 import { Category } from '../core/entities'
 
 export class CategoryService implements CategoryRepository {
   constructor(private readonly workspaceService: WorkspaceRepository) {}
+
+  private async categoyBelongsToWorkspace(
+    workspaceId: string,
+    categoryId: number
+  ): Promise<CategoriesDto | undefined> {
+    const categories = await this.getByWorkspace(workspaceId)
+
+    const categoryBelongsToWorkspace = (categories as CategoriesDto[]).find(
+      (category) => category.id === categoryId
+    )
+
+    return categoryBelongsToWorkspace
+  }
 
   async getByWorkspace(id: string): Promise<CategoriesDto[] | CustomError> {
     const existWorkspace = await this.workspaceService.get(id)
@@ -75,17 +98,69 @@ export class CategoryService implements CategoryRepository {
     }
   }
 
-  async delete(id: number): Promise<Category | CustomError> {
+  async update(
+    id: number,
+    data: UpdateCategoryDto
+  ): Promise<Category | CustomError> {
     if (isNaN(id)) {
       return new BadRequest('El id de la categoría debe ser un número')
     }
 
+    const exist = await this.get(id)
+
+    if (exist instanceof CustomError) {
+      return exist
+    }
+
+    const canUpdate = await this.categoyBelongsToWorkspace(data.workspaceId, id)
+
+    if (!canUpdate) {
+      return new NotAuthorized(
+        'El usuario no está autorizado para editar una categoría de otro workspace'
+      )
+    }
+
+    try {
+      const updatedCategory = await prisma.category.update({
+        where: { id },
+        data,
+      })
+      return updatedCategory
+    } catch (error) {
+      console.log(error)
+      return new Conflict(`Error al actualizar la categoría`)
+    }
+  }
+
+  async delete(
+    id: number,
+    workspaceId: string
+  ): Promise<Category | CustomError> {
+    if (isNaN(id)) {
+      return new BadRequest('El id de la categoría debe ser un número')
+    }
+
+    const exist = await this.get(id)
+
+    if (exist instanceof CustomError) {
+      return exist
+    }
+
+    const canDelete = await this.categoyBelongsToWorkspace(workspaceId, id)
+
+    if (!canDelete) {
+      return new NotAuthorized(
+        'El usuario no está autorizado para borrar una categoría de otro workspace'
+      )
+    }
+
     try {
       const deletedCategory = await prisma.category.delete({ where: { id } })
+
       return deletedCategory
     } catch (error) {
       console.log(error)
-      return new NotFound('Categoría no encontrada')
+      return new Conflict(`Error al eliminar la categoría`)
     }
   }
 }
